+++ FinEdgeFrontend/src/Hooks/Transactions/useGetCategorySankeyData.tsx
@@ -1,12 +1,12 @@
 import { useQuery } from "@tanstack/react-query";
 import axios, { AxiosResponse, AxiosError } from "axios";
-import { getCategorySankeyChartDataEndPoint } from "../../endpoints";
+import { getTransactionSankeyChartDataEndPoint } from "../../endpoints";
 import { SankeyData } from "@progress/kendo-react-all";
 
-const useGetSankeyChartData = () => {
+const useGetTransactionSankeyChartData = () => {
   const getSankeyChartData = async () => {
     return await axios
-      .get<SankeyData>(`${getCategorySankeyChartDataEndPoint}`, { withCredentials: true, })
+      .get<SankeyData>(`${getTransactionSankeyChartDataEndPoint}`, { withCredentials: true, })
       .then((res: AxiosResponse<SankeyData>) => res.data)
       .catch((err: AxiosError) => {
         throw new Error(`Failed to fetch Sankey chart data: ${err.message}`);
@@ -23,4 +23,4 @@ const useGetSankeyChartData = () => {
   return { data, isLoading, isError, error, };
 };
 
-export default useGetSankeyChartData;
+export default useGetTransactionSankeyChartData;
