+++ FinEdgeFrontend/src/Pages/Home.tsx
@@ -1,7 +1,7 @@
 import InfoCard from "../Components/HomePage/Cards/InfoCard";
 import StatCard from "../Components/HomePage/Cards/StatCard";
 import UserProfile from "../Components/HomePage/Cards/UserProfile";
-import useGetReports from "../Hooks/useGetReports";
+import useGetReports from "../Hooks/Transactions/useGetReports";
 import useGetUser from "../Hooks/useGetUser";
 import AccountsGrid from "../Components/HomePage/Grids/AccountsGrid";
 import CategoriesGrid from "../Components/HomePage/Grids/CategoriesGrid";
